# Conflicts:
#	bp-dao/src/main/java/com/x/bp/dao/po/ProductDO.java

+++ bp-core/src/main/java/com/x/bp/core/dto/order/OrderQueryReq.java
@@ -11,7 +11,7 @@ public class OrderQueryReq extends BaseQuery {
     /**
      * 平台
      */
-    private Integer platform;
+    private Long platform;
 
     /**
      * 用户ID
