+++ bp-dao/src/main/java/com/x/bp/dao/po/OrderDO.java
@@ -32,7 +32,7 @@ public class OrderDO {
 
     private Long userId;
 
-    private Integer platform;
+    private Long platform;
 
     private Integer orderStatus;
